(UPDATE) Out of matches display
If you are out of matches and visit the matching page, it loads for 20 seconds before telling you youre out of matches.
This kind of opens up the case where the query takes more than 20 seconds but there ARE matches but thats pretty unlikely.
Maybe theres a better way to do this.